refactor: rename ternary pkg to condition
